Single Source Shortest Path Algorithm
Dijkstra's algorithm is an greedy algorithm used for finding the shortest paths between nodes in a graph, which may represent, for example, road networks.
The algorithm exists in many variants. Dijkstra's original algorithm found the shortest path between two given nodes, but a more common variant fixes a single node as the "source" node and finds shortest paths from the source to all other nodes in the graph, producing a shortest-path tree.

For a given source node in the graph, the algorithm finds the shortest path between that node and every other. It can also be used for finding the shortest paths from a single node to a single destination node by stopping the algorithm once the shortest path to the destination node has been determined. For example, if the nodes of the graph represent places in a city and edge path costs represent distances between pairs of places connected by a direct road, Dijkstra's algorithm can be used to find the shortest route between one place and all other places.